Throw early if SWIG is not installed
